Update and rename 3.2.1.1.c to 3.2.11.c

+++ 3.2.11.c
@@ -1,8 +1,8 @@
 #include <stdio.h>
 #include <stdlib.h>
-double* foo(int n)
+int* foo(int n)
 {
-    return malloc(n*sizeof(double));
+    return malloc(n*sizeof(int));
 }
 int main()
 {
